% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constituency}
\alias{constituency}
\title{Danish constituencies}
\format{A Large SpatialPolygonsDataFrame}
\usage{
constituency
}
\description{
A set of spatial polygons to form danish constituencies
}
\keyword{datasets}

