% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencyDK.R
\name{constituencyDK}
\alias{constituencyDK}
\title{Creates dynamic chloropleth maps of danish constituencies}
\usage{
constituencyDK(value = NULL, id = NULL, subplot = NULL, data = NULL,
  map = FALSE, legend = FALSE)
}
\arguments{
\item{value}{is the name of the column in the data where the values are stored}

\item{id}{is the name of the column in the data where the ids are stored}

\item{subplot}{is a string of ids you want to keep - excludes all others}

\item{data}{is the data frame that contains the data to map}

\item{map}{is a TRUE / FALSE of wether a real map should be plotted underneath}
}
\value{
An interactive Leaflet map
}
\description{
Creates dynamic chloropleth maps of danish constituencies
}
\examples{
library(leafletDK)
library(dplyr)
library(tidyr)
library(stringr)

fv15tot <- read.csv2("http://api.statbank.dk/v1/data/FV15TOT/CSV?VALRES=VAELG\%2C13&OMR\%C3\%85DE=*",
                     stringsAsFactors = FALSE, encoding = "UTF-8")

fv15tot <- fv15tot \%>\% spread(VALRES, INDHOLD) \%>\% select(-TID)
names(fv15tot) <- c("area", "writing", "voters")
fv15tot <- fv15tot \%>\% mutate(pertenthousand = writing/(voters/10000))
fv15tot <- fv15tot \%>\% filter(str_detect(fv15tot$area, "STORKREDS")) \%>\%
  mutate(area = str_to_lower(str_replace_all(area, "S STORKREDS", "")))

constituencyDK("pertenthousand", "area", data = fv15tot)

}

