% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district}
\alias{district}
\title{Danish districts}
\format{A Large SpatialPolygonsDataFrame}
\usage{
district
}
\description{
A set of spatial polygons to form danish nominating districts
}
\keyword{datasets}

