% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs.R
\name{getIDs}
\alias{getIDs}
\title{Gives you the ids of a given area}
\usage{
getIDs(area)
}
\arguments{
\item{area}{is the name of the area you want ids from. must be either:
\itemize{
  \item "constituency"
  \item "district"
  \item "municipal"
  \item "parish"
  \item "regional"
  \item "rural"
  \item "zip"
}}
}
\value{
String of ids for the given area
}
\description{
Gives you the ids of a given area
}
\examples{
library(leafletDK)

getIDs("constituency")

}

