% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{join_map_data_zip}
\alias{join_map_data_zip}
\title{Helper function}
\usage{
join_map_data_zip(value = NULL, id = NULL, subplot = NULL,
  mapdata = NULL, shapefile)
}
\description{
Helper function
}

