% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipal}
\alias{municipal}
\title{Danish municipalities}
\format{A Large SpatialPolygonsDataFrame}
\usage{
municipal
}
\description{
A set of spatial polygons to form danish municipalities
}
\keyword{datasets}

