% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parishDK.R
\name{parishDK}
\alias{parishDK}
\title{Creates dynamic chloropleth maps of danish parishes}
\usage{
parishDK(value = NULL, id = NULL, subplot = NULL, data = NULL,
  map = FALSE, legend = FALSE, pal = "YlOrRd", logcol = F)
}
\arguments{
\item{value}{is the name of the column in the data where the values are stored}

\item{id}{is the name of the column in the data where the ids are stored}

\item{subplot}{is a string of ids you want to keep - excludes all others}

\item{data}{is the data frame that contains the data to map}

\item{map}{is a TRUE / FALSE of wether a real map should be plotted underneath}
}
\value{
An interactive Leaflet map
}
\description{
Creates dynamic chloropleth maps of danish parishes
}
\examples{
library(leafletDK)

km1 <- readr::read_csv2("http://api.statbank.dk/v1/data/KM1/CSV?SOGN=*&FKMED=*")
km1 <- tidyr::spread(km1, FKMED, INDHOLD)
km1$SOGN <- stringr::str_trim(stringr::str_replace_all(km1$SOGN, "(\\\\d+)|([-])", ""))
km1 <- dplyr::group_by(km1, SOGN)
km1 <- dplyr::summarise(km1,
                        medlem = sum(`Medlem af Folkekirken`, na.rm = T),
                        ikke_medlem = sum(`Ikke medlem af Folkekirken`, na.rm = T))
km1 <- dplyr::mutate(km1, pct = medlem/(ikke_medlem + medlem), pct = round(pct*100,2))

parishDK("pct", "SOGN", data = km1)

}

