% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruralDK.R
\name{ruralDK}
\alias{ruralDK}
\title{Creates dynamic chloropleth maps of danish rural regions}
\usage{
ruralDK(value = NULL, id = NULL, subplot = NULL, data = NULL)
}
\arguments{
\item{value}{is the name of the column in the data where the values are stored}

\item{id}{is the name of the column in the data where the ids are stored}

\item{subplot}{is a string of ids you want to keep - excludes all others}

\item{data}{is the data frame that contains the data to map}
}
\value{
An interactive Leaflet map
}
\description{
Creates dynamic chloropleth maps of danish rural regions
}
\examples{
library(leafletDK)
aus08 <- read.csv2("http://api.statbank.dk/v1/data/AUS08/CSV?OMR\%C3\%85DE=01\%2C02\%2C03\%2C04\%2C05\%2C06\%2C07\%2C08\%2C09\%2C10\%2C11&SAESONFAK=9",
                   stringsAsFactors = F)
aus08$OMRÅDE <- str_replace_all(aus08$OMRÅDE, "Landsdel ", "")

ruralDK("INDHOLD", "OMRÅDE", data = aus08)

}

