% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipDK.R
\name{zipDK}
\alias{zipDK}
\title{Creates dynamic chloropleth maps of danish zip codes}
\usage{
zipDK(value = NULL, id = NULL, subplot = NULL, data = NULL,
  map = FALSE, legend = FALSE, pal = "YlOrRd", logcol = F,
  legendtitle = NULL)
}
\arguments{
\item{value}{is the name of the column in the data where the values are stored}

\item{id}{is the name of the column in the data where the ids are stored}

\item{subplot}{is a string of ids you want to keep - excludes all others}

\item{data}{is the data frame that contains the data to map}

\item{map}{is a TRUE / FALSE of wether a real map should be plotted underneath}

\item{legend}{is a TRUE / FALSE of wether the legend should be plotted}

\item{pal}{is the color palette for the chloropleth}

\item{logcol}{is a TRUE / FALSE of wether the chloropleth colors should be drawn against the log value}

\item{legendtitle}{provides an alternative title for the legend}
}
\value{
An interactive Leaflet map
}
\description{
Creates dynamic chloropleth maps of danish zip codes
}
\examples{
library(leafletDK)

ejen4 <- readr::read_csv2("http://api.statbank.dk/v1/data/EJEN4/CSV?OMR\%C3\%85DE=*&EJENDOMSKATE=2103&BN\%C3\%98GLE=2&OVERDRAG=1&Tid=2001")

ejen4 <- dplyr::select(ejen4, OMRÅDE, TID, INDHOLD)
ejen4 <- ejen4[stringr::str_detect(ejen4$OMRÅDE,"[A-ZÆØÅa-zæøå]"),]
ejen4$zip <- unlist(stringr::str_extract_all(ejen4$OMRÅDE, "\\\\d{4}"))
ejen4$zip <- ifelse(as.numeric(ejen4$zip) < 2400, round(as.numeric(ejen4$zip)/100)*100, as.numeric(ejen4$zip))
ejen4 <- dplyr::group_by(ejen4, zip)
ejen4 <- dplyr::summarise(ejen4, indhold = sum(INDHOLD))

zipDK("indhold", "zip", data = ejen4)

}

