% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parishDK.R
\name{parishDK}
\alias{parishDK}
\title{Creates dynamic chloropleth maps of danish parishes}
\usage{
parishDK(value = NULL, id = NULL, subplot = NULL, data = NULL)
}
\arguments{
\item{value}{is the name of the column in the data where the values are stored}

\item{id}{is the name of the column in the data where the ids are stored}

\item{subplot}{is a string of ids you want to keep - excludes all others}

\item{data}{is the data frame that contains the data to map}
}
\value{
An interactive Leaflet map
}
\description{
Creates dynamic chloropleth maps of danish parishes
}
\examples{
library(leafletDK)
library(tidyr)

km1 <- read.csv2("http://api.statbank.dk/v1/data/KM1/CSV?SOGN=*&FKMED=*",
                 stringsAsFactors = F)
km1 <- km1 \%>\% spread(FKMED, INDHOLD)
km1$pct <- round(km1[,4]/(km1[,3]+km1[,4])*100,1)

parishDK("pct", "SOGN", data = km1)

}

