% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalDK.R
\name{regionalDK}
\alias{regionalDK}
\title{Creates dynamic chloropleth maps of danish regions}
\usage{
regionalDK(value = NULL, id = NULL, subplot = NULL, data = NULL,
  map = F, legend = F)
}
\arguments{
\item{value}{is the name of the column in the data where the values are stored}

\item{id}{is the name of the column in the data where the ids are stored}

\item{subplot}{is a string of ids you want to keep - excludes all others}

\item{data}{is the data frame that contains the data to map}

\item{map}{is a TRUE / FALSE of wether a real map should be plotted underneath}
}
\value{
An interactive Leaflet map
}
\description{
Creates dynamic chloropleth maps of danish regions
}
\examples{
library(leafletDK)

folk1 <- read.csv2("http://api.statbank.dk/v1/data/folk1/CSV?OMR\%C3\%85DE=*",
                     stringsAsFactors = F)

regionalDK("INDHOLD", "OMRÅDE", subplot = c("frederiksberg","hvidovre"), data = folk1)

}

