% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipDK.R
\name{zipDK}
\alias{zipDK}
\title{Creates dynamic chloropleth maps of danish zip codes}
\usage{
zipDK(value = NULL, id = NULL, subplot = NULL, data = NULL, map = F,
  legend = F)
}
\arguments{
\item{value}{is the name of the column in the data where the values are stored}

\item{id}{is the name of the column in the data where the ids are stored}

\item{subplot}{is a string of ids you want to keep - excludes all others}

\item{data}{is the data frame that contains the data to map}

\item{map}{is a TRUE / FALSE of wether a real map should be plotted underneath}
}
\value{
An interactive Leaflet map
}
\description{
Creates dynamic chloropleth maps of danish zip codes
}
\examples{
library(leafletDK)
library(dplyr)
library(tidyr)
library(stringr)

ejen4 <- read.csv2("http://api.statbank.dk/v1/data/EJEN4/CSV?OMR\%C3\%85DE=*&EJENDOMSKATE=2103&BN\%C3\%98GLE=2&OVERDRAG=1&Tid=2001",
                   stringsAsFactors = F)

ejen4 <- ejen4 \%>\% select(OMRÅDE, TID, INDHOLD)
ejen4 <- ejen4[str_detect(ejen4$OMRÅDE,"[A-ZÆØÅa-zæøå]"),]
ejen4$zip <- unlist(str_extract_all(ejen4$OMRÅDE, "\\\\d{4}"))
ejen4$zip <- ifelse(as.numeric(ejen4$zip) < 2400, round(as.numeric(ejen4$zip)/100)*100, as.numeric(ejen4$zip))
ejen4 <- ejen4 \%>\% group_by(zip) \%>\% summarise(indhold = sum(INDHOLD))
ejen4 <- as.data.frame(ejen4)

zipDK("indhold", "zip", data = ejen4)

}

